FileStore/CreateException is FilePath/CreateException
