+++ src/Exception/FilePath/NonAbsolutePathException.php
@@ -4,17 +4,12 @@ declare(strict_types=1);
 
 namespace App\Exception\FilePath;
 
-class NonAbsolutePathException extends \Exception implements FilePathExceptionInterface
+class NonAbsolutePathException extends AbstractFilePathException
 {
     public function __construct(
-        private string $path,
+        string $path,
         ?\Throwable $previous = null
     ) {
-        parent::__construct(sprintf('Path "%s" is not absolute"', $path), 0, $previous);
-    }
-
-    public function getPath(): string
-    {
-        return $this->path;
+        parent::__construct($path, sprintf('Path "%s" is not absolute"', $path), $previous);
     }
 }
